added few testcases to param_tests file and casted ft_strlen to int in parameter_format function

+++ sources/parameter_formatting/parameter_format.c
@@ -6,7 +6,7 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 11:40:05 by mviinika          #+#    #+#             */
-/*   Updated: 2023/02/21 21:25:37 by mviinika         ###   ########.fr       */
+/*   Updated: 2023/02/21 21:48:12 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -104,7 +104,7 @@ int	param_format(char **cmd)
 			ret = proceed_to_execute(&pa, &ints, cmd, &ret);
 			if (ret == -1)
 				return (-1);
-			if (ints.j > ft_strlen(cmd[ints.i]))
+			if (ints.j > (int)ft_strlen(cmd[ints.i]))
 				break ;
 		}
 		ints.j = 0;
